Reorganise files and exports
